A quick guide for the pictures
This version is simply a guide for how I think the pictures should be.
They, off course, look terrible and require an external package to
generate them, but just wrote this to make sure we could talk about the
file structure for where to store the images and generate some talk
about them in general.

Also, kivy is always flaky, so looking for non-kivy things to do :]